+++ packages/types-core/src/interfaces.ts
@@ -3,7 +3,7 @@
 
 export type DynamicDatasourceCreator = (name: string, args: Record<string, unknown>) => Promise<void>;
 
-export interface Cache {
-  set<D>(key: string, value: D): Promise<void>;
-  get<D>(key: string): Promise<D | undefined>;
+export interface Cache<T extends Record<string, any> = Record<string, any>> {
+  set(key: keyof T, value: T[keyof T]): Promise<void>;
+  get(key: keyof T): Promise<T[keyof T] | undefined>;
 }