+++ packages/node-core/CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ## Added
-- Add in-memory cache to be used in projects #(2110)
+- Add in-memory cache to be used in sandbox #(2110)
 
 ## [6.0.3] - 2023-10-17
 ### Fixed